fix: responsive variations not working
